JAVAFIXTURE-27: Merge ParameterizedType-implementation into TypeCapture
Keeping OpenParameterizedType as it's own class for just one use-case
seems unnecessary. Also the wobbly class-name sounded wrong from the
beginning. So I removed it and use TypeCapture to create an anonymous
object from the interface ParameterizedType.
